app/requirements.txt
Django==4.1.1
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
python-dotenv==0.21.0
martor==1.6.14
gunicorn==20.1.0
psycopg2-binary==2.9.1
